+++ packages/atomic/src/components/atomic-pager/atomic-pager.stories.tsx
@@ -1,11 +1,9 @@
 import defaultStory from '../../../.storybook/default-story';
-import PagerDoc from './atomic-pager.mdx';
 
 const {defaultModuleExport, exportedStory} = defaultStory(
   'Atomic/Pager',
   'atomic-pager',
-  {},
-  PagerDoc
+  {}
 );
 
 export default defaultModuleExport;