+++ packages/atomic/src/components/facets/atomic-category-facet/atomic-category-facet.stories.tsx
@@ -1,12 +1,10 @@
 import {html} from 'lit-html';
 import defaultStory from '../../../../.storybook/default-story';
-import CategoryFacetDoc from './atomic-category-facet.mdx';
 
 const {defaultModuleExport, exportedStory} = defaultStory(
   'Atomic/CategoryFacet',
   'atomic-category-facet',
   {field: 'geographicalhierarchy'},
-  CategoryFacetDoc,
   {
     additionalMarkup: () => html`<style>
       atomic-category-facet {