+++ packages/atomic/src/components/facets/atomic-rating-facet/atomic-rating-facet.stories.tsx
@@ -1,12 +1,10 @@
 import {html} from 'lit-html';
 import defaultStory from '../../../../.storybook/default-story';
-import RatingFacetDoc from './atomic-rating-facet.mdx';
 
 const {defaultModuleExport, exportedStory} = defaultStory(
   'Atomic/RatingFacet',
   'atomic-rating-facet',
   {field: 'snrating'},
-  RatingFacetDoc,
   {
     additionalMarkup: () => html`<style>
       atomic-rating-facet {