+++ packages/atomic/stencil.config.ts
@@ -61,11 +61,12 @@ export const config: Config = {
       copy: [
         {src: 'pages', keepDirStructure: false},
         {src: 'themes'},
+        isDevWatch && {src: 'external-builds', dest: 'build/headless'},
         {
           src: '../node_modules/@salesforce-ux/design-system/assets/icons/{doctype,standard}/*.svg',
           dest: 'build/assets',
         },
-      ],
+      ].filter((n) => n),
     },
   ],
   testing: {
@@ -76,6 +77,8 @@ export const config: Config = {
     },
     transformIgnorePatterns: [],
     testPathIgnorePatterns: ['headless'],
+    setupFiles: ['jest-localstorage-mock'],
+    resetMocks: false,
   },
   devServer: {
     reloadStrategy: 'pageReload',