+++ packages/headless/src/case-assist.index.ts
@@ -24,3 +24,37 @@ export {
 export {LoggerOptions} from './app/logger';
 
 export {LogLevel} from './app/logger';
+
+// Case Assist Action Loaders
+export * from './features/case-input/case-input-actions-loader';
+export * from './features/case-field/case-field-actions-loader';
+export * from './features/document-suggestion/document-suggestion-actions-loader';
+export * from './features/case-assist/case-assist-analytics-actions-loader';
+
+// Controllers
+export {
+  Controller,
+  buildController,
+} from './controllers/controller/headless-controller';
+
+export {
+  CaseInputState,
+  CaseInput,
+  CaseInputOptions,
+  CaseInputProps,
+  buildCaseInput,
+} from './controllers/case-input/headless-case-input';
+
+export {
+  CaseFieldState,
+  CaseField,
+  CaseFieldOptions,
+  CaseFieldProps,
+  buildCaseField,
+} from './controllers/case-field/headless-case-field';
+
+export {
+  DocumentSuggestionList as DocumentSuggestion,
+  DocumentSuggestionListState as DocumentSuggestionState,
+  buildDocumentSuggestionList as buildDocumentSuggestion,
+} from './controllers/document-suggestion-list/headless-document-suggestion-list';